requirements.txt
anthropic>=0.7.0
python-dotenv>=1.0.0
eth-account>=0.10.0
aiohttp>=3.9.0
openai>=1.0.0
agents>=0.1.0  # For OpenAI Agents SDK
web3>=6.0.0    # For Ethereum interactions 